completed user profile

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video"
+import User from "../models/User"
 
 export const trending = async(req,res)=> {
 	const videos = await Video.find({});
@@ -7,7 +8,7 @@ export const trending = async(req,res)=> {
 
 export const watch = async(req,res) => {
     const {id}= req.params;
-	const video = await Video.findById(id);
+	const video = await Video.findById(id).populate("owner");
     if(video===null) {
         return res.status(404).render("404", {pageTitle: "Video Not Found"});
     }
@@ -16,19 +17,27 @@ export const watch = async(req,res) => {
 
 export const getEdit =async(req,res) => {
     const {id}= req.params;
+    const {loggedInUser : {_id}}= req.session;
     const video = await Video.findById(id);
     if(video===null) {
         return res.render("404", {pageTitle: "Video Not Found"});
+    }
+    if (String(video.owner) !== String(_id)) { 
+        return res.status(403).redirect("/");
     }
     return res.render("edit",{pageTitle:`Editing : ${video.title}`, video});
 };
 
 export const postEdit = async(req,res) => {
     const {id}= req.params;
+    const {loggedInUser : {_id}}= req.session;
     const {title, description, hashtags} = req.body;
     const video = await Video.findById(id);
     if(video===null) {
         return res.render("404", {pageTitle: "Video Not Found"});
+    }
+    if (String(video.owner) !== String(_id)) { 
+        return res.status(403).redirect("/");
     }
     try { 
         await Video.findByIdAndUpdate(id, {title, description, hashtags:Video.hashtagFormatting(hashtags)})
@@ -43,13 +52,18 @@ export const getUpload = (req, res) => {
 }; 
 
 export const postUpload = async (req, res) => {
-    const {title, description, hashtags} = req.body;
+    const { session: {loggedInUser: {_id} }, file: {path} , body: {title, description, hashtags}} = req;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             hashtags: Video.hashtagFormatting(hashtags),
+            fileUrl:path,
+            owner: _id
         });
+        const user = await User.findById(_id);
+        await user.videos.push(newVideo._id);
+        await user.save();
         return res.redirect("/");
     } catch(error) {  
         return res.render("upload", { pageTitle: "Upload Video", errorMessage: error._message,})
@@ -58,6 +72,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req,res) => {
     const {id}=req.params;
+    const {loggedInUser : {_id}}= req.session;
+    const video = await Video.findById(id);
+    if(!video) {
+        return res.render("404", {pageTitle: "Video Not Found"});
+    }
+    if (String(video.owner) !== String(_id)) { 
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
@@ -70,5 +92,3 @@ export const search = async(req,res) => {
 	 }
     res.render("search",{pageTitle:"Search Video", videos});
 } 
-
-export const upload =(req,res) => res.render("upload");
